Signed-off-by: ricardo cedeño montaña <drnn1076@linux-9xum.site>
General upgrade of the program to Processing 2.0

Major changes in the XML functions and font functions.

Label.pde. This is a new file. It stores all the bibliographic information about the graphic.
Before that information was given in the main sketch.

The font used in the sketch is now included too as .ttf file